testing/20140401.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20140401</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20140401.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2014-04-01 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Tue Apr 01 00:00:00 2014</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3" _a="heycam|away"><a href="#l-3">#</a> [00:04] * heycam|away is now known as heycam</li>
 <li id="l-4"><a href="#l-4">#</a> [00:05] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-5"><a href="#l-5">#</a> [00:10] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-6"><a href="#l-6">#</a> [00:11] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;nn&quot;)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:17] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-8"><a href="#l-8">#</a> [00:20] * Quits: edoyle (~edoyle@public.cloak) (&quot;&quot;)</li>
 <li id="l-9"><a href="#l-9">#</a> [00:21] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-10"><a href="#l-10">#</a> [00:35] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-11"><a href="#l-11">#</a> [00:43] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-12"><a href="#l-12">#</a> [00:43] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-13"><a href="#l-13">#</a> [01:03] * Quits: lukeis (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-14" _a="heycam"><a href="#l-14">#</a> [01:05] * heycam is now known as heycam|away</li>
 <li id="l-15" _a="heycam|away"><a href="#l-15">#</a> [01:11] * heycam|away is now known as heycam</li>
 <li id="l-16"><a href="#l-16">#</a> [01:24] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-17"><a href="#l-17">#</a> [01:37] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-18"><a href="#l-18">#</a> [01:37] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-19"><a href="#l-19">#</a> [01:44] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-20"><a href="#l-20">#</a> [02:02] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-21"><a href="#l-21">#</a> [02:07] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-22"><a href="#l-22">#</a> [02:14] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-23"><a href="#l-23">#</a> [02:28] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-24"><a href="#l-24">#</a> [02:28] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-25"><a href="#l-25">#</a> [02:35] * Quits: rhauck (~Adium@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-26"><a href="#l-26">#</a> [02:45] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-27"><a href="#l-27">#</a> [03:08] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-28" _a="heycam"><a href="#l-28">#</a> [03:11] * heycam is now known as heycam|away</li>
 <li id="l-29"><a href="#l-29">#</a> [03:15] * Joins: yankhates_cb (~yankhatescb@public.cloak)</li>
 <li id="l-30"><a href="#l-30">#</a> [03:15] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-31"><a href="#l-31">#</a> [03:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [03:22] -gitbot:#testing- [web-platform-tests] yunxiaoxie opened pull request #824: Move and improve idlharness tests for IndexedDB (master...yunxiaoxie/interfaces) <a href="https://github.com/w3c/web-platform-tests/pull/824">https://github.com/w3c/web-platform-tests/pull/824</a></li>
 <li id="l-33"><a href="#l-33">#</a> [03:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-34"><a href="#l-34">#</a> [03:29] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-35"><a href="#l-35">#</a> [03:29] -gitbot:#testing- [web-platform-tests] zqzhang pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/a30b7fd57ed6f1a0207b3c06f5ce7eedb1c636a9">https://github.com/w3c/web-platform-tests/commit/a30b7fd57ed6f1a0207b3c06f5ce7eedb1c636a9</a></li>
 <li id="l-36"><a href="#l-36">#</a> [03:29] -gitbot:#testing- web-platform-tests/master a30b7fd Odin Hørthe Omdal: IndexedDB: Delete the Opera-ported Microsoft IDB tests...</li>
 <li id="l-37"><a href="#l-37">#</a> [03:29] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-38"><a href="#l-38">#</a> [03:31] * Quits: yankhates_cb (~yankhatescb@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-39"><a href="#l-39">#</a> [03:33] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-40"><a href="#l-40">#</a> [03:33] -gitbot:#testing- [web-platform-tests] zqzhang closed pull request #516: Move ported Opera IndexedDB tests (master...Velmont/move_ported_idb_tests) <a href="https://github.com/w3c/web-platform-tests/pull/516">https://github.com/w3c/web-platform-tests/pull/516</a></li>
 <li id="l-41"><a href="#l-41">#</a> [03:33] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-42"><a href="#l-42">#</a> [03:33] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-43"><a href="#l-43">#</a> [03:33] -gitbot:#testing- [web-platform-tests] zqzhang deleted Velmont/move_ported_idb_tests at 1a9d933: <a href="https://github.com/w3c/web-platform-tests/commit/1a9d933">https://github.com/w3c/web-platform-tests/commit/1a9d933</a></li>
 <li id="l-44"><a href="#l-44">#</a> [03:33] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-45"><a href="#l-45">#</a> [03:35] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-46"><a href="#l-46">#</a> [03:35] -gitbot:#testing- [web-platform-tests] zqzhang closed pull request #818: Remove test files in IndexedDB/submissions/Opera/invalid (master...yunxiaoxie/invalid) <a href="https://github.com/w3c/web-platform-tests/pull/818">https://github.com/w3c/web-platform-tests/pull/818</a></li>
 <li id="l-47"><a href="#l-47">#</a> [03:35] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-48"><a href="#l-48">#</a> [03:37] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-49"><a href="#l-49">#</a> [03:44] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-50"><a href="#l-50">#</a> [03:44] -gitbot:#testing- [web-platform-tests] zqzhang pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/0f38898942f851d952c36834e974d6121ef35193">https://github.com/w3c/web-platform-tests/commit/0f38898942f851d952c36834e974d6121ef35193</a></li>
 <li id="l-51"><a href="#l-51">#</a> [03:44] -gitbot:#testing- web-platform-tests/master 0f38898 Xie Yunxiao: Move a test for createObjectStore to IndexedDB/</li>
 <li id="l-52"><a href="#l-52">#</a> [03:44] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-53"><a href="#l-53">#</a> [03:45] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-54"><a href="#l-54">#</a> [03:45] -gitbot:#testing- [web-platform-tests] zqzhang closed pull request #819: Move a test for createObjectStore to IndexedDB/ (master...yunxiaoxie/createObjectStore) <a href="https://github.com/w3c/web-platform-tests/pull/819">https://github.com/w3c/web-platform-tests/pull/819</a></li>
 <li id="l-55"><a href="#l-55">#</a> [03:45] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-56"><a href="#l-56">#</a> [04:09] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-57"><a href="#l-57">#</a> [04:16] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-58" _a="heycam|away"><a href="#l-58">#</a> [04:16] * heycam|away is now known as heycam</li>
 <li id="l-59"><a href="#l-59">#</a> [05:10] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-60"><a href="#l-60">#</a> [05:17] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-61"><a href="#l-61">#</a> [06:11] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-62"><a href="#l-62">#</a> [06:18] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-63"><a href="#l-63">#</a> [06:34] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-64"><a href="#l-64">#</a> [06:34] -gitbot:#testing- [web-platform-tests] shishimaru opened pull request #825: Tests for URL.createObjectURL(mediaSource) in Media Source Extensions. (master...shishimaru/URLObjectExtensions-createObjectURL) <a href="https://github.com/w3c/web-platform-tests/pull/825">https://github.com/w3c/web-platform-tests/pull/825</a></li>
 <li id="l-65"><a href="#l-65">#</a> [06:34] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-66"><a href="#l-66">#</a> [06:38] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-67"><a href="#l-67">#</a> [06:52] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-68"><a href="#l-68">#</a> [07:11] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-69"><a href="#l-69">#</a> [07:19] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-70"><a href="#l-70">#</a> [07:36] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-71"><a href="#l-71">#</a> [08:10] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-72"><a href="#l-72">#</a> [08:10] -gitbot:#testing- [web-platform-tests] hallvors force-pushed hallvors/XHRbugFix3 from d80d91f to 33b771b: <a href="https://github.com/w3c/web-platform-tests/commits/hallvors/XHRbugFix3">https://github.com/w3c/web-platform-tests/commits/hallvors/XHRbugFix3</a></li>
 <li id="l-73"><a href="#l-73">#</a> [08:10] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 29a5ef7 Hallvord R. M. Steen: call abort() from loadstart, otherwise test is invalid</li>
 <li id="l-74"><a href="#l-74">#</a> [08:10] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 a817dcd Hallvord R. M. Steen: minor bug fix</li>
 <li id="l-75"><a href="#l-75">#</a> [08:10] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 612db5c Hallvord R. M. Steen: improved test coverage of abort() algorithm 4.3 and 5</li>
 <li id="l-76"><a href="#l-76">#</a> [08:10] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-77"><a href="#l-77">#</a> [08:12] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-78"><a href="#l-78">#</a> [08:19] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-79"><a href="#l-79">#</a> [08:20] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-80"><a href="#l-80">#</a> [08:20] -gitbot:#testing- [web-platform-tests] hallvors pushed 1 new commit to hallvors/XHRbugFix3: <a href="https://github.com/w3c/web-platform-tests/commit/45dfeddc5e7a962a8d68ebe44fa5832dd9b0aa84">https://github.com/w3c/web-platform-tests/commit/45dfeddc5e7a962a8d68ebe44fa5832dd9b0aa84</a></li>
 <li id="l-81"><a href="#l-81">#</a> [08:20] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 45dfedd Hallvord R. M. Steen: not sure how this .orig file got added..</li>
 <li id="l-82"><a href="#l-82">#</a> [08:20] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-83"><a href="#l-83">#</a> [08:21] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-84"><a href="#l-84">#</a> [08:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-85"><a href="#l-85">#</a> [08:22] -gitbot:#testing- [web-platform-tests] hallvors pushed 1 new commit to hallvors/XHRbugFix3: <a href="https://github.com/w3c/web-platform-tests/commit/d1df7df7fa1b8f32b0dad0120e540b70e580f3fc">https://github.com/w3c/web-platform-tests/commit/d1df7df7fa1b8f32b0dad0120e540b70e580f3fc</a></li>
 <li id="l-86"><a href="#l-86">#</a> [08:22] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 d1df7df Hallvord R. M. Steen: drop hopefully obsolete .php file</li>
 <li id="l-87"><a href="#l-87">#</a> [08:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-88" _a="heycam"><a href="#l-88">#</a> [08:34] * heycam is now known as heycam|away</li>
 <li id="l-89"><a href="#l-89">#</a> [09:02] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-90"><a href="#l-90">#</a> [09:02] -gitbot:#testing- [web-platform-tests] yunxiaoxie opened pull request #826: Add more tests for IDBCursor advance (master...yunxiaoxie/idbcursor-advance) <a href="https://github.com/w3c/web-platform-tests/pull/826">https://github.com/w3c/web-platform-tests/pull/826</a></li>
 <li id="l-91"><a href="#l-91">#</a> [09:02] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-92"><a href="#l-92">#</a> [09:13] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-93"><a href="#l-93">#</a> [09:20] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-94"><a href="#l-94">#</a> [09:22] * Joins: shepazu (schepers@public.cloak)</li>
 <li id="l-95"><a href="#l-95">#</a> [09:23] * Quits: shepazu (schepers@public.cloak) (&quot;is probably traveling...&quot;)</li>
 <li id="l-96"><a href="#l-96">#</a> [09:34] * Joins: yankhates_cb (~yankhatescb@public.cloak)</li>
 <li id="l-97"><a href="#l-97">#</a> [09:50] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-98"><a href="#l-98">#</a> [09:52] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-99"><a href="#l-99">#</a> [10:04] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-100" _a="zcorpan" _r="jgraham"><a href="#l-100">#</a> [10:10] &lt;zcorpan&gt; jgraham: i've received the chocolate now. cheers!</li>
 <li id="l-101"><a href="#l-101">#</a> [10:16] * Quits: ptressel (~chatzilla@public.cloak) (Client closed connection)</li>
 <li id="l-102"><a href="#l-102">#</a> [10:17] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-103"><a href="#l-103">#</a> [10:27] * Joins: anssik (~uid10742@public.cloak)</li>
 <li id="l-104"><a href="#l-104">#</a> [10:37] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-105" _a="jgraham" _r="zcorpan"><a href="#l-105">#</a> [10:39] &lt;jgraham&gt; zcorpan: np, you deserve it! :)</li>
 <li id="l-106" _a="darobin" _r="jgraham"><a href="#l-106">#</a> [10:45] &lt;darobin&gt; jgraham: did you know that window.opener was set when you open a link through &lt;a target=_blank&gt;?</li>
 <li id="l-107" _a="darobin" _r="jgraham"><a href="#l-107">#</a> [10:46] &lt;darobin&gt; jgraham: head to <a href="http://w3c.github.io/dom/test-results/less-than-2.html">http://w3c.github.io/dom/test-results/less-than-2.html</a>, and click on one of the bold blue-on-grey links in the table</li>
 <li id="l-108" _a="darobin" _r="Error"><a href="#l-108">#</a> [10:47] &lt;darobin&gt; Error: Permission denied to access property 'testharness_properties'</li>
 <li id="l-109" _a="darobin" _r="jgraham"><a href="#l-109">#</a> [10:50] &lt;darobin&gt; jgraham: unrelatedly, in the manifest generation code you're using git status --porcelain --ignore-submodules=untracked to check that the repo is clean</li>
 <li id="l-110" _a="darobin"><a href="#l-110">#</a> [10:50] &lt;darobin&gt; but that doesn't exclude dirty submodules (irrespective of ignore: dirty)</li>
 <li id="l-111" _a="darobin"><a href="#l-111">#</a> [10:51] &lt;darobin&gt; is that intentional?</li>
 <li id="l-112" _a="darobin"><a href="#l-112">#</a> [10:51] &lt;darobin&gt; I can fix if not</li>
 <li id="l-113"><a href="#l-113">#</a> [10:56] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-114"><a href="#l-114">#</a> [10:56] -gitbot:#testing- [web-platform-tests] darobin pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/228703a87299c3742caa31eea39c5c75bebeefe3">https://github.com/w3c/web-platform-tests/commit/228703a87299c3742caa31eea39c5c75bebeefe3</a></li>
 <li id="l-115"><a href="#l-115">#</a> [10:56] -gitbot:#testing- web-platform-tests/master 228703a Robin Berjon: updated resources submodules with latest changes to TH</li>
 <li id="l-116"><a href="#l-116">#</a> [10:56] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-117" _a="jgraham" _r="darobin"><a href="#l-117">#</a> [10:57] &lt;jgraham&gt; darobin: Not intentional</li>
 <li id="l-118" _a="darobin"><a href="#l-118">#</a> [10:59] &lt;darobin&gt; done</li>
 <li id="l-119"><a href="#l-119">#</a> [10:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-120"><a href="#l-120">#</a> [10:59] -gitbot:#testing- [web-platform-tests] darobin pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/e3a6a575ae3fc8064026debc7401d235a0a5a1a2">https://github.com/w3c/web-platform-tests/commit/e3a6a575ae3fc8064026debc7401d235a0a5a1a2</a></li>
 <li id="l-121"><a href="#l-121">#</a> [10:59] -gitbot:#testing- web-platform-tests/master e3a6a57 Robin Berjon: make manifest generator ignore all changes in submodules as they shouldn't matter</li>
 <li id="l-122"><a href="#l-122">#</a> [10:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-123" _a="darobin" _r="jgraham"><a href="#l-123">#</a> [11:01] &lt;darobin&gt; jgraham: <a href="https://github.com/w3c/testharness.js">https://github.com/w3c/testharness.js</a> is the right TH nowadays right?</li>
 <li id="l-124"><a href="#l-124">#</a> [11:06] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-125" _a="darobin"><a href="#l-125">#</a> [11:10] &lt;darobin&gt; is there a prescribed level of browser support for testharness?</li>
 <li id="l-126" _a="darobin"><a href="#l-126">#</a> [11:11] * darobin wondering if he can ditch IE8, FF3.6, SF4, OP11.5, Konq4.3</li>
 <li id="l-127" _a="darobin" _r="jgraham"><a href="#l-127">#</a> [11:13] &lt;darobin&gt; jgraham: in <a href="https://github.com/w3c/testharness.js/blob/master/testharnessreport.js#L176">https://github.com/w3c/testharness.js/blob/master/testharnessreport.js#L176</a> you're using a variable that's defined in the else block 19 lines down — is that your intent?</li>
 <li id="l-128" _a="zcorpan"><a href="#l-128">#</a> [11:21] &lt;zcorpan&gt; does web driver support selecting a file for &lt;input type=file&gt;?</li>
 <li id="l-129" _a="darobin" _r="jgraham"><a href="#l-129">#</a> [11:25] &lt;darobin&gt; jgraham: on lines <a href="https://github.com/w3c/testharness.js/blob/master/testharness.js#L1467">https://github.com/w3c/testharness.js/blob/master/testharness.js#L1467</a> and 1507 what is self supposed to be? it never seems defined</li>
 <li id="l-130" _a="jgraham" _r="darobin"><a href="#l-130">#</a> [11:25] &lt;jgraham&gt; darobin: You keep using this word &quot;you&quot;. I don't think it means what you think it means ;)</li>
 <li id="l-131" _a="jgraham"><a href="#l-131">#</a> [11:25] &lt;jgraham&gt; (I didn't write the thr.js code)</li>
 <li id="l-132" _a="darobin"><a href="#l-132">#</a> [11:26] &lt;darobin&gt; oh!</li>
 <li id="l-133" _a="darobin" _r="sorry"><a href="#l-133">#</a> [11:26] &lt;darobin&gt; sorry, I thought it was you</li>
 <li id="l-134" _a="darobin"><a href="#l-134">#</a> [11:26] &lt;darobin&gt; that said the last bit is th.js</li>
 <li id="l-135" _a="jgraham"><a href="#l-135">#</a> [11:27] &lt;jgraham&gt; <a href="http://software.hixie.ch/utilities/js/live-dom-viewer/saved/2911">http://software.hixie.ch/utilities/js/live-dom-viewer/saved/2911</a></li>
 <li id="l-136" _a="jgraham" _r="darobin"><a href="#l-136">#</a> [11:27] &lt;jgraham&gt; darobin: That's the right repo, yes</li>
 <li id="l-137" _a="jgraham"><a href="#l-137">#</a> [11:27] &lt;jgraham&gt; I don't think there</li>
 <li id="l-138" _a="darobin" _r="oh"><a href="#l-138">#</a> [11:27] &lt;darobin&gt; oh, that self</li>
 <li id="l-139" _a="jgraham"><a href="#l-139">#</a> [11:27] &lt;jgraham&gt; 's a prescribed level of browser support</li>
 <li id="l-140" _a="jgraham"><a href="#l-140">#</a> [11:28] * jgraham marks it down as &quot;not self-evident&quot;</li>
 <li id="l-141" _a="jgraham"><a href="#l-141">#</a> [11:28] &lt;jgraham&gt; (sorry)</li>
 <li id="l-142" _a="darobin"><a href="#l-142">#</a> [11:28] &lt;darobin&gt; people will often use var self = this when they need to refer to this in callback and such</li>
 <li id="l-143" _a="jgraham"><a href="#l-143">#</a> [11:28] &lt;jgraham&gt; Yeah</li>
 <li id="l-144" _a="darobin"><a href="#l-144">#</a> [11:28] &lt;darobin&gt; I'll make that window.self, easier</li>
 <li id="l-145" _a="darobin"><a href="#l-145">#</a> [11:28] * darobin is doing a jshint pass since it kept giving errors</li>
 <li id="l-146"><a href="#l-146">#</a> [11:32] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-147"><a href="#l-147">#</a> [11:34] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-148"><a href="#l-148">#</a> [11:34] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #802: import tests on document.open() from presto-testo (master...open-input-stream) <a href="https://github.com/w3c/web-platform-tests/pull/802">https://github.com/w3c/web-platform-tests/pull/802</a></li>
 <li id="l-149"><a href="#l-149">#</a> [11:34] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-150"><a href="#l-150">#</a> [11:38] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-151"><a href="#l-151">#</a> [11:38] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #761: tests on input type password (master...input-password) <a href="https://github.com/w3c/web-platform-tests/pull/761">https://github.com/w3c/web-platform-tests/pull/761</a></li>
 <li id="l-152"><a href="#l-152">#</a> [11:38] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-153"><a href="#l-153">#</a> [11:41] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-154"><a href="#l-154">#</a> [11:41] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #763: add tests on input state change (master...input-change-state) <a href="https://github.com/w3c/web-platform-tests/pull/763">https://github.com/w3c/web-platform-tests/pull/763</a></li>
 <li id="l-155"><a href="#l-155">#</a> [11:41] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-156"><a href="#l-156">#</a> [11:41] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-157"><a href="#l-157">#</a> [11:41] -gitbot:#testing- [web-platform-tests] zcorpan pushed 4 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/3132c3e462f4...7600f0a00412">https://github.com/w3c/web-platform-tests/compare/3132c3e462f4...7600f0a00412</a></li>
 <li id="l-158"><a href="#l-158">#</a> [11:41] -gitbot:#testing- web-platform-tests/master 411e096 Denis Ah-Kang: add tests on input state change</li>
 <li id="l-159"><a href="#l-159">#</a> [11:41] -gitbot:#testing- web-platform-tests/master fb0cb4e Denis Ah-Kang: fix description and test</li>
 <li id="l-160"><a href="#l-160">#</a> [11:41] -gitbot:#testing- web-platform-tests/master 95d6522 Denis Ah-Kang: add test on input value mode</li>
 <li id="l-161"><a href="#l-161">#</a> [11:41] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-162"><a href="#l-162">#</a> [12:04] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-163"><a href="#l-163">#</a> [12:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-164"><a href="#l-164">#</a> [12:22] -gitbot:#testing- [web-platform-tests] zcorpan pushed 2 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/7600f0a00412...458767f22d31">https://github.com/w3c/web-platform-tests/compare/7600f0a00412...458767f22d31</a></li>
 <li id="l-165"><a href="#l-165">#</a> [12:22] -gitbot:#testing- web-platform-tests/master dd7705c plehegar: Yet another s/php/py/</li>
 <li id="l-166"><a href="#l-166">#</a> [12:22] -gitbot:#testing- web-platform-tests/master 458767f Simon Pieters: Merge pull request #816 from plehegar/plh/404-3...</li>
 <li id="l-167"><a href="#l-167">#</a> [12:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-168"><a href="#l-168">#</a> [12:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-169"><a href="#l-169">#</a> [12:22] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #816: Yet another s/php/py/ (master...plh/404-3) <a href="https://github.com/w3c/web-platform-tests/pull/816">https://github.com/w3c/web-platform-tests/pull/816</a></li>
 <li id="l-170"><a href="#l-170">#</a> [12:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-171" _a="darobin"><a href="#l-171">#</a> [12:26] &lt;darobin&gt; I collated the full DOM4 results for five browsers with a nice (hopefully) stylesheet: <a href="http://w3c.github.io/dom/test-results/all.html">http://w3c.github.io/dom/test-results/all.html</a></li>
 <li id="l-172" _a="darobin"><a href="#l-172">#</a> [12:27] &lt;darobin&gt; that said I need to make it better for bz who apparently uses a small browser window</li>
 <li id="l-173"><a href="#l-173">#</a> [12:30] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-174" _a="abarsto"><a href="#l-174">#</a> [12:30] * abarsto is now known as ArtB</li>
 <li id="l-175" _a="Ms2ger"><a href="#l-175">#</a> [12:33] &lt;Ms2ger&gt; bz has lots of windows</li>
 <li id="l-176" _a="darobin"><a href="#l-176">#</a> [12:48] &lt;darobin&gt; and lots of wisdoms</li>
 <li id="l-177" _a="darobin" _r="ooooh"><a href="#l-177">#</a> [12:48] &lt;darobin&gt; ooooh, who's the funny person who gave a test name of &quot;constructor&quot;?</li>
 <li id="l-178" _a="darobin"><a href="#l-178">#</a> [12:48] &lt;darobin&gt; that does fun things when you try to play with the data structure</li>
 <li id="l-179" _a="darobin"><a href="#l-179">#</a> [12:48] &lt;darobin&gt; #wtfjs</li>
 <li id="l-180" _a="Ms2ger"><a href="#l-180">#</a> [12:49] &lt;Ms2ger&gt; Probably me</li>
 <li id="l-181" _a="Ms2ger"><a href="#l-181">#</a> [12:49] &lt;Ms2ger&gt; Use a Map!</li>
 <li id="l-182" _a="Ms2ger" _r="Wow"><a href="#l-182">#</a> [12:52] &lt;Ms2ger&gt; Wow, failures on &quot;For Element nodes, nodeName should return the same as tagName.&quot;?</li>
 <li id="l-183" _a="darobin"><a href="#l-183">#</a> [12:52] &lt;darobin&gt; some of those things are suspect</li>
 <li id="l-184" _a="darobin"><a href="#l-184">#</a> [12:52] &lt;darobin&gt; heh</li>
 <li id="l-185" _a="darobin"><a href="#l-185">#</a> [12:52] &lt;darobin&gt; I guess I could run with --harmony_collections</li>
 <li id="l-186"><a href="#l-186">#</a> [12:54] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-187"><a href="#l-187">#</a> [12:55] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-188"><a href="#l-188">#</a> [12:55] -gitbot:#testing- [web-platform-tests] zcorpan pushed 5 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/458767f22d31...41a3aefffa2e">https://github.com/w3c/web-platform-tests/compare/458767f22d31...41a3aefffa2e</a></li>
 <li id="l-189"><a href="#l-189">#</a> [12:55] -gitbot:#testing- web-platform-tests/master f1de0a9 xiaojunwu: Add tests for the RadioNodeList interface</li>
 <li id="l-190"><a href="#l-190">#</a> [12:55] -gitbot:#testing- web-platform-tests/master 8c4fbbb xiaojunwu: Add tests for the HTMLFormControlsCollection interface</li>
 <li id="l-191"><a href="#l-191">#</a> [12:55] -gitbot:#testing- web-platform-tests/master 8c23b93 xiaojunwu: Update according to review comments</li>
 <li id="l-192"><a href="#l-192">#</a> [12:55] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-193"><a href="#l-193">#</a> [13:07] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-194"><a href="#l-194">#</a> [13:07] -gitbot:#testing- [web-platform-tests] zcorpan pushed 3 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/41a3aefffa2e...a75a86120742">https://github.com/w3c/web-platform-tests/compare/41a3aefffa2e...a75a86120742</a></li>
 <li id="l-195"><a href="#l-195">#</a> [13:07] -gitbot:#testing- web-platform-tests/master 07aa9ae plehegar: Worker origin needs to take into account server port number</li>
 <li id="l-196"><a href="#l-196">#</a> [13:07] -gitbot:#testing- web-platform-tests/master cde4185 plehegar: Dealt with empty location.port</li>
 <li id="l-197"><a href="#l-197">#</a> [13:07] -gitbot:#testing- web-platform-tests/master a75a861 Simon Pieters: Merge pull request #817 from plehegar/plh/worker-origin-fix...</li>
 <li id="l-198"><a href="#l-198">#</a> [13:07] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-199" _a="Ms2ger" _r="assert_array_equals"><a href="#l-199">#</a> [13:11] &lt;Ms2ger&gt; assert_array_equals: property 1, expected Element node &lt;abc&gt;&lt;/abc&gt; but got Element node &lt;abc&gt;&lt;/abc&gt;</li>
 <li id="l-200" _a="jgraham"><a href="#l-200">#</a> [13:12] &lt;jgraham&gt; Different nodes?</li>
 <li id="l-201" _a="Ms2ger" _r="Yeah"><a href="#l-201">#</a> [13:13] &lt;Ms2ger&gt; Yeah, probably</li>
 <li id="l-202" _a="zcorpan"><a href="#l-202">#</a> [13:14] * zcorpan stashes <a href="https://critic.hoppipolla.co.uk/showcomment?chain=2839">https://critic.hoppipolla.co.uk/showcomment?chain=2839</a></li>
 <li id="l-203" _a="darobin" _r="mmmmm"><a href="#l-203">#</a> [13:15] &lt;darobin&gt; mmmmm, so for some reason using Opera 12.16, just getting <a href="http://my-test-setup/MANIFEST.json">http://my-test-setup/MANIFEST.json</a> always yields something like <a href="https://gist.github.com/anonymous/03b472962a1ad8e1c43c">https://gist.github.com/anonymous/03b472962a1ad8e1c43c</a></li>
 <li id="l-204" _a="Ms2ger" _r="jgraham"><a href="#l-204">#</a> [13:15] &lt;Ms2ger&gt; jgraham, fwiw, to get the parent sha, I tend to just trawl through git log</li>
 <li id="l-205" _a="darobin"><a href="#l-205">#</a> [13:16] &lt;darobin&gt; and it seems to do the same thing against w3c-test.org</li>
 <li id="l-206" _a="darobin"><a href="#l-206">#</a> [13:16] * darobin gives up on getting Presto results for the HTML TS for now</li>
 <li id="l-207" _a="Ms2ger" _r="Oh"><a href="#l-207">#</a> [13:17] &lt;Ms2ger&gt; Oh, are you fucking kidding me</li>
 <li id="l-208" _a="Ms2ger"><a href="#l-208">#</a> [13:17] &lt;Ms2ger&gt;                 case Node.ELEMENT_NODE:</li>
 <li id="l-209" _a="Ms2ger"><a href="#l-209">#</a> [13:17] &lt;Ms2ger&gt;                     var ret = &quot;&lt;&quot; + val.tagName.toLowerCase();</li>
 <li id="l-210" _a="jgraham" _r="Ms2ger"><a href="#l-210">#</a> [13:18] &lt;jgraham&gt; Ms2ger: Well yeah, I often get it by doing &quot;fetch branch&quot; in critic, clicking on the parent commit, ignoring the data loss warnings, copying the SHA1 and doing Fetch Branch again with the right upstream. But I wouldn't recommend that to anyone else :)</li>
 <li id="l-211" _a="Ms2ger"><a href="#l-211">#</a> [13:18] &lt;Ms2ger&gt; Turns out it actually means &quot;expected Element node &lt;abc&gt;&lt;/abc&gt; but got Element node &lt;Abc&gt;&lt;/Abc&gt;&quot;</li>
 <li id="l-212" _a="jgraham" _r="Oh"><a href="#l-212">#</a> [13:19] &lt;jgraham&gt; Oh, that seems confusing</li>
 <li id="l-213" _a="jgraham" _r="Patch"><a href="#l-213">#</a> [13:19] &lt;jgraham&gt; Patch, please</li>
 <li id="l-214" _a="Ms2ger"><a href="#l-214">#</a> [13:19] &lt;Ms2ger&gt; Yeah</li>
 <li id="l-215" _a="jgraham" _r="darobin"><a href="#l-215">#</a> [13:19] &lt;jgraham&gt; darobin: Seems like Opera's closing the connection early, maybe?</li>
 <li id="l-216" _a="jgraham"><a href="#l-216">#</a> [13:19] &lt;jgraham&gt; But it doesn't happen to me</li>
 <li id="l-217" _a="Ms2ger"><a href="#l-217">#</a> [13:20] &lt;Ms2ger&gt; Found two bugs in a test of Aryeh's, yay</li>
 <li id="l-218" _a="Ms2ger"><a href="#l-218">#</a> [13:20] &lt;Ms2ger&gt; And one in his th.js code</li>
 <li id="l-219" _a="darobin" _r="jgraham"><a href="#l-219">#</a> [13:22] &lt;darobin&gt; jgraham: yes, but I can't figure out why it would do that</li>
 <li id="l-220" _a="darobin"><a href="#l-220">#</a> [13:22] &lt;darobin&gt; I've set it to use no cache, to no avail</li>
 <li id="l-221" _a="jgraham"><a href="#l-221">#</a> [13:22] * jgraham finds it is lunch time, hopes not to discover any bugs in that</li>
 <li id="l-222" _a="darobin"><a href="#l-222">#</a> [13:22] &lt;darobin&gt; when it gets JSON directly it pops up a download dialog and I wonder if that's when it stops downloading</li>
 <li id="l-223" _a="darobin"><a href="#l-223">#</a> [13:23] &lt;darobin&gt; but I can't see to get it to just show me the JSON</li>
 <li id="l-224" _a="darobin"><a href="#l-224">#</a> [13:23] &lt;darobin&gt; *seem</li>
 <li id="l-225" _a="Ms2ger" _r="jgraham"><a href="#l-225">#</a> [13:24] &lt;Ms2ger&gt; jgraham, r? <a href="https://github.com/w3c/testharness.js/pull/66">https://github.com/w3c/testharness.js/pull/66</a></li>
 <li id="l-226" _a="darobin" _r="anyway"><a href="#l-226">#</a> [13:24] &lt;darobin&gt; anyway, I'll run Safari instead</li>
 <li id="l-227"><a href="#l-227">#</a> [13:26] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-228" _a="Ms2ger"><a href="#l-228">#</a> [13:26] &lt;Ms2ger&gt; Also r? <a href="https://github.com/w3c/web-platform-tests/pull/827">https://github.com/w3c/web-platform-tests/pull/827</a>, but after the previous one</li>
 <li id="l-229"><a href="#l-229">#</a> [13:26] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-230"><a href="#l-230">#</a> [13:26] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #827: Improve and correct the getElementsByTagName tests in case.js. (master...case-getElementsByTagName) <a href="https://github.com/w3c/web-platform-tests/pull/827">https://github.com/w3c/web-platform-tests/pull/827</a></li>
 <li id="l-231"><a href="#l-231">#</a> [13:26] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-232" _a="Ms2ger" _r="darobin"><a href="#l-232">#</a> [13:27] &lt;Ms2ger&gt; darobin, sorry, messed with your test results :)</li>
 <li id="l-233"><a href="#l-233">#</a> [13:29] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-234" _a="Ms2ger" _r="darobin"><a href="#l-234">#</a> [13:29] &lt;Ms2ger&gt; darobin, oh, and Fx's excellent score on Range tests is definitely foul play... We've fixed bugs the tests uncovered ;)</li>
 <li id="l-235"><a href="#l-235">#</a> [13:31] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-236"><a href="#l-236">#</a> [13:34] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-237"><a href="#l-237">#</a> [13:37] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-238" _a="darobin" _r="Ms2ger"><a href="#l-238">#</a> [13:39] &lt;darobin&gt; Ms2ger: hehehe</li>
 <li id="l-239" _a="darobin"><a href="#l-239">#</a> [13:39] &lt;darobin&gt; how did you mess with my results?</li>
 <li id="l-240"><a href="#l-240">#</a> [13:39] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-241" _a="Ms2ger" _r="darobin"><a href="#l-241">#</a> [13:39] &lt;Ms2ger&gt; darobin, the above PR changes the results</li>
 <li id="l-242" _a="darobin" _r="Ms2ger"><a href="#l-242">#</a> [13:39] &lt;darobin&gt; Ms2ger: oh, that's fine, I expect that</li>
 <li id="l-243" _a="darobin"><a href="#l-243">#</a> [13:40] &lt;darobin&gt; do you know if it changes them a lot?</li>
 <li id="l-244" _a="darobin"><a href="#l-244">#</a> [13:40] &lt;darobin&gt; doesn't look bad</li>
 <li id="l-245" _a="Ms2ger"><a href="#l-245">#</a> [13:40] &lt;Ms2ger&gt; Changes five lines and adds five more, I think</li>
 <li id="l-246" _a="darobin" _r="yeah"><a href="#l-246">#</a> [13:41] &lt;darobin&gt; yeah, that's not a big deal :)</li>
 <li id="l-247" _a="darobin"><a href="#l-247">#</a> [13:41] &lt;darobin&gt; the results are preliminary anyway</li>
 <li id="l-248" _a="Ms2ger"><a href="#l-248">#</a> [13:41] &lt;Ms2ger&gt; Can you do something to highlight the tests that fail in all browsers?</li>
 <li id="l-249" _a="Ms2ger"><a href="#l-249">#</a> [13:42] &lt;Ms2ger&gt; &quot;First set attribute is returned by setAttribute&quot;</li>
 <li id="l-250"><a href="#l-250">#</a> [13:42] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-251" _a="Ms2ger"><a href="#l-251">#</a> [13:42] &lt;Ms2ger&gt; I think I'm going to deny I wrote that</li>
 <li id="l-252" _a="Ms2ger"><a href="#l-252">#</a> [13:44] &lt;Ms2ger&gt; And maybe add the test failure message as a tooltip, if that's easy</li>
 <li id="l-253"><a href="#l-253">#</a> [13:45] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-254" _a="Ms2ger"><a href="#l-254">#</a> [13:46] &lt;Ms2ger&gt; Hah</li>
 <li id="l-255" _a="Ms2ger" _r="Fwiw"><a href="#l-255">#</a> [13:46] &lt;Ms2ger&gt; Fwiw, the test with that summary was:</li>
 <li id="l-256" _a="Ms2ger"><a href="#l-256">#</a> [13:46] &lt;Ms2ger&gt;   el.setAttributeNS(&quot;ab&quot;, &quot;attr&quot;, &quot;t1&quot;)</li>
 <li id="l-257" _a="Ms2ger"><a href="#l-257">#</a> [13:46] &lt;Ms2ger&gt;   el.setAttributeNS(&quot;kl&quot;, &quot;attr&quot;, &quot;t2&quot;)</li>
 <li id="l-258" _a="Ms2ger"><a href="#l-258">#</a> [13:46] &lt;Ms2ger&gt;   assert_equals(el.getAttribute(&quot;attr&quot;), &quot;t2&quot;)</li>
 <li id="l-259"><a href="#l-259">#</a> [13:47] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-260"><a href="#l-260">#</a> [13:47] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #828: Fix typos in attributes.html. (master...attributes) <a href="https://github.com/w3c/web-platform-tests/pull/828">https://github.com/w3c/web-platform-tests/pull/828</a></li>
 <li id="l-261"><a href="#l-261">#</a> [13:47] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-262"><a href="#l-262">#</a> [13:48] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-263"><a href="#l-263">#</a> [13:49] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-264"><a href="#l-264">#</a> [13:49] * Quits: Lachy (~Lachy@public.cloak) (Client closed connection)</li>
 <li id="l-265"><a href="#l-265">#</a> [13:49] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-266"><a href="#l-266">#</a> [14:01] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-267" _a="darobin"><a href="#l-267">#</a> [14:48] * darobin wonders if Servo has test results</li>
 <li id="l-268" _a="darobin"><a href="#l-268">#</a> [14:48] &lt;darobin&gt; at this point the JSON I got out of my WebKit run shows it passing 5% of the HTML TS — I wonder if Servo does better :)</li>
 <li id="l-269" _a="jgraham"><a href="#l-269">#</a> [14:49] &lt;jgraham&gt; I doubt it. It doesn't even implement range</li>
 <li id="l-270" _a="jgraham"><a href="#l-270">#</a> [14:49] &lt;jgraham&gt; :p</li>
 <li id="l-271" _a="darobin"><a href="#l-271">#</a> [14:49] &lt;darobin&gt; this is HTML :)</li>
 <li id="l-272" _a="jgraham" _r="Oh"><a href="#l-272">#</a> [14:49] &lt;jgraham&gt; Oh, fair point</li>
 <li id="l-273" _a="jgraham"><a href="#l-273">#</a> [14:49] &lt;jgraham&gt; Speaking of servo</li>
 <li id="l-274" _a="jgraham" _r="Ms2ger"><a href="#l-274">#</a> [14:49] &lt;jgraham&gt; Ms2ger: innerHTML review?</li>
 <li id="l-275" _a="jgraham" _r="Oh"><a href="#l-275">#</a> [14:51] &lt;jgraham&gt; Oh, maybe zcorpan did it already</li>
 <li id="l-276" _a="zcorpan"><a href="#l-276">#</a> [14:51] &lt;zcorpan&gt; which one?</li>
 <li id="l-277" _a="Ms2ger"><a href="#l-277">#</a> [14:52] &lt;Ms2ger&gt; I would certainly be happy if he did :)</li>
 <li id="l-278" _a="jgraham"><a href="#l-278">#</a> [14:52] &lt;jgraham&gt; r/523</li>
 <li id="l-279" _a="jgraham"><a href="#l-279">#</a> [14:52] &lt;jgraham&gt; Looks like all the issues are in the tests that I didn't write :(</li>
 <li id="l-280" _a="Ms2ger"><a href="#l-280">#</a> [14:52] &lt;Ms2ger&gt; In tests that I wrote, then?</li>
 <li id="l-281" _a="Ms2ger"><a href="#l-281">#</a> [14:53] &lt;Ms2ger&gt; I can probably look at the comments later</li>
 <li id="l-282" _a="jgraham" _r="Dunno"><a href="#l-282">#</a> [14:53] &lt;jgraham&gt; Dunno, did you write the other tests there?</li>
 <li id="l-283" _a="jgraham"><a href="#l-283">#</a> [14:54] &lt;jgraham&gt; Anyway we should land it soon so that darobin cries^Whas useful results for his DOM IR</li>
 <li id="l-284" _a="Ms2ger"><a href="#l-284">#</a> [14:54] &lt;Ms2ger&gt; I suspect yes</li>
 <li id="l-285" _a="Ms2ger"><a href="#l-285">#</a> [14:54] &lt;Ms2ger&gt; Is he putting innerHTML in his fork?</li>
 <li id="l-286" _a="darobin"><a href="#l-286">#</a> [14:55] &lt;darobin&gt; only on my spoon</li>
 <li id="l-287"><a href="#l-287">#</a> [14:58] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-288"><a href="#l-288">#</a> [15:06] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-289"><a href="#l-289">#</a> [15:06] -gitbot:#testing- [web-platform-tests] hallvors pushed 3 new commits to hallvors/XHRbugFix3: <a href="https://github.com/w3c/web-platform-tests/compare/d1df7df7fa1b...fde5297e01bb">https://github.com/w3c/web-platform-tests/compare/d1df7df7fa1b...fde5297e01bb</a></li>
 <li id="l-290"><a href="#l-290">#</a> [15:06] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 6b6b4b2 Hallvord R. M. Steen: replacing some new .php helper files with .py ones</li>
 <li id="l-291"><a href="#l-291">#</a> [15:06] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 fb89202 Hallvord R. M. Steen: bugfixing</li>
 <li id="l-292"><a href="#l-292">#</a> [15:06] -gitbot:#testing- web-platform-tests/hallvors/XHRbugFix3 fde5297 Hallvord R. M. Steen: minor bugfixes</li>
 <li id="l-293"><a href="#l-293">#</a> [15:06] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-294" _a="sangwhan"><a href="#l-294">#</a> [15:31] &lt;sangwhan&gt; someone asked me if w-p-t has a coverage graph/table of some sort, my answer was no but i figured someone here can prove me wrong</li>
 <li id="l-295" _a="gsnedders" _r="jgraham"><a href="#l-295">#</a> [15:32] &lt;gsnedders&gt; jgraham: Ms2ger : if you can look athe html5lib reviews I pinged soon that would be nice. Might manage to do something in the next week before exam time hits.</li>
 <li id="l-296" _a="gsnedders" _r="sangwhan"><a href="#l-296">#</a> [15:33] &lt;gsnedders&gt; sangwhan: Coverage of the specs?</li>
 <li id="l-297" _a="Ms2ger" _r="gsnedders"><a href="#l-297">#</a> [15:34] &lt;Ms2ger&gt; gsnedders, happy to defer to jgraham :)</li>
 <li id="l-298" _a="gsnedders" _r="Ms2ger"><a href="#l-298">#</a> [15:34] &lt;gsnedders&gt; Ms2ger: I don't care who reviews them! :P</li>
 <li id="l-299" _a="Ms2ger"><a href="#l-299">#</a> [15:34] &lt;Ms2ger&gt; I do :)</li>
 <li id="l-300" _a="gsnedders"><a href="#l-300">#</a> [15:35] &lt;gsnedders&gt; I presume you too have exams coming up?</li>
 <li id="l-301" _a="Ms2ger"><a href="#l-301">#</a> [15:35] &lt;Ms2ger&gt; Not just yet, fortunately</li>
 <li id="l-302" _a="Ms2ger"><a href="#l-302">#</a> [15:36] &lt;Ms2ger&gt; But enough projects to keep me occupied</li>
 <li id="l-303" _a="sangwhan" _r="gsnedders"><a href="#l-303">#</a> [15:36] &lt;sangwhan&gt; gsnedders: coverage of specs, yes</li>
 <li id="l-304" _a="gsnedders"><a href="#l-304">#</a> [15:36] * gsnedders is on his week &quot;holiday&quot; between submitting what is effectively BSci thesis and exam revision now, so doing stuff he mostly wants to</li>
 <li id="l-305" _a="gsnedders"><a href="#l-305">#</a> [15:37] &lt;gsnedders&gt; I'd put the thesis online somewhere, but it's pretty useless. Essentially concludes, &quot;I spent my entire time learning about formal methods and as a result proved very little&quot;.</li>
 <li id="l-306" _a="gsnedders"><a href="#l-306">#</a> [15:38] &lt;gsnedders&gt; (&quot;and I'm not sure my model is correct anyway&quot;)</li>
 <li id="l-307" _a="Ms2ger"><a href="#l-307">#</a> [15:39] &lt;Ms2ger&gt; I'd still like to read it :)</li>
 <li id="l-308" _a="jgraham"><a href="#l-308">#</a> [15:40] * jgraham is not exactly lacking things to do :(</li>
 <li id="l-309" _a="gsnedders"><a href="#l-309">#</a> [15:40] &lt;gsnedders&gt; Ping me later when I'm not on a tethered 2G connection and I might upload it. :)</li>
 <li id="l-310" _a="Ms2ger"><a href="#l-310">#</a> [15:40] * Ms2ger needs to think of a subject for his masters thesis at some point</li>
 <li id="l-311" _a="gsnedders"><a href="#l-311">#</a> [15:41] &lt;gsnedders&gt; How big of a project? I can give you my list for ~800 hour projects. :P</li>
 <li id="l-312" _a="Ms2ger"><a href="#l-312">#</a> [15:42] &lt;Ms2ger&gt; Ha</li>
 <li id="l-313" _a="Ms2ger"><a href="#l-313">#</a> [15:43] &lt;Ms2ger&gt; Too short, sorry, I need 840 hours :)</li>
 <li id="l-314" _a="gsnedders" _r="Ms2ger"><a href="#l-314">#</a> [15:47] &lt;gsnedders&gt; Ms2ger: If you want something directly applicable, try and come up with some model based on n-grams and TLD to determine character encoding in HTML, so we can get rid o the locale-speciic defaults?</li>
 <li id="l-315" _a="sangwhan"><a href="#l-315">#</a> [15:48] &lt;sangwhan&gt; wonder if doing a brute force search for calls to th.js based on IDL extracted from specs in web-platform-tests would give a usable overview of what is covered</li>
 <li id="l-316" _a="gsnedders"><a href="#l-316">#</a> [15:51] &lt;gsnedders&gt; It's not a useful view of coverage though, IMO. Because tests for each branch in the algorithms is way more important.</li>
 <li id="l-317" _a="gsnedders" _r="Like"><a href="#l-317">#</a> [15:51] &lt;gsnedders&gt; Like, merely knowing we have one test per IDL line is pretty irrelevant, really.</li>
 <li id="l-318" _a="astearns" _r="gsnedders"><a href="#l-318">#</a> [16:03] &lt;astearns&gt; gsnedders: do you think it's &gt;=1? knowing there are IDL lines with no tests is pretty useful</li>
 <li id="l-319" _a="gsnedders"><a href="#l-319">#</a> [16:04] &lt;gsnedders&gt; If you include idlharness.js stuff, it's pretty much guaranteed to be &gt;=1</li>
 <li id="l-320" _a="gsnedders" _r="bbiab"><a href="#l-320">#</a> [16:04] &lt;gsnedders&gt; bbiab, changing trains</li>
 <li id="l-321"><a href="#l-321">#</a> [16:07] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-322"><a href="#l-322">#</a> [16:11] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-323"><a href="#l-323">#</a> [16:28] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-324"><a href="#l-324">#</a> [16:36] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-325"><a href="#l-325">#</a> [16:36] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-326"><a href="#l-326">#</a> [16:54] * Joins: mdyck (~mdyck@public.cloak)</li>
 <li id="l-327"><a href="#l-327">#</a> [17:03] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-328"><a href="#l-328">#</a> [17:04] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-329"><a href="#l-329">#</a> [17:11] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-330"><a href="#l-330">#</a> [17:20] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-331"><a href="#l-331">#</a> [17:23] * Joins: lukeis (~Adium@public.cloak)</li>
 <li id="l-332" _a="sangwhan"><a href="#l-332">#</a> [17:27] * sangwhan sense gsnedders asking for work :)</li>
 <li id="l-333" _a="Ms2ger" _r="sangwhan"><a href="#l-333">#</a> [17:36] &lt;Ms2ger&gt; sangwhan, r? <a href="https://critic.hoppipolla.co.uk/r/1145">https://critic.hoppipolla.co.uk/r/1145</a> :)</li>
 <li id="l-334"><a href="#l-334">#</a> [17:41] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-335" _a="sangwhan" _r="Ms2ger"><a href="#l-335">#</a> [17:41] &lt;sangwhan&gt; Ms2ger: took a quick look, don't think i'm the right person to sign off on that :)</li>
 <li id="l-336" _a="Ms2ger"><a href="#l-336">#</a> [17:42] &lt;Ms2ger&gt; Dammit :)</li>
 <li id="l-337" _a="Ms2ger"><a href="#l-337">#</a> [17:42] &lt;Ms2ger&gt; gsnedders?</li>
 <li id="l-338"><a href="#l-338">#</a> [17:42] * Quits: yankhates_cb (~yankhatescb@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-339"><a href="#l-339">#</a> [17:53] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-340"><a href="#l-340">#</a> [17:53] -gitbot:#testing- [web-platform-tests] jgraham pushed 2 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/a75a86120742...e86e6fe1768b">https://github.com/w3c/web-platform-tests/compare/a75a86120742...e86e6fe1768b</a></li>
 <li id="l-341"><a href="#l-341">#</a> [17:53] -gitbot:#testing- web-platform-tests/master 2a84de2 Ms2ger: Improve and correct the getElementsByTagName tests in case.js.</li>
 <li id="l-342"><a href="#l-342">#</a> [17:53] -gitbot:#testing- web-platform-tests/master e86e6fe jgraham: Merge pull request #827 from Ms2ger/case-getElementsByTagName...</li>
 <li id="l-343"><a href="#l-343">#</a> [17:53] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-344" _a="Ms2ger"><a href="#l-344">#</a> [17:53] &lt;Ms2ger&gt; Next up, <a href="https://github.com/w3c/web-platform-tests/pull/828">https://github.com/w3c/web-platform-tests/pull/828</a></li>
 <li id="l-345" _a="plh"><a href="#l-345">#</a> [17:56] &lt;plh&gt; that one I was able to do :)</li>
 <li id="l-346" _a="Ms2ger"><a href="#l-346">#</a> [17:58] &lt;Ms2ger&gt; Yay :)</li>
 <li id="l-347"><a href="#l-347">#</a> [17:58] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-348"><a href="#l-348">#</a> [17:58] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 2 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/e86e6fe1768b...07e7ef04288b">https://github.com/w3c/web-platform-tests/compare/e86e6fe1768b...07e7ef04288b</a></li>
 <li id="l-349"><a href="#l-349">#</a> [17:58] -gitbot:#testing- web-platform-tests/master eb65ce5 Ms2ger: Fix typos in attributes.html.</li>
 <li id="l-350"><a href="#l-350">#</a> [17:58] -gitbot:#testing- web-platform-tests/master 07e7ef0 Ms2ger: Merge pull request #828 from Ms2ger/attributes...</li>
 <li id="l-351"><a href="#l-351">#</a> [17:58] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-352" _a="gsnedders" _r="sangwhan"><a href="#l-352">#</a> [18:03] &lt;gsnedders&gt; sangwhan: bah, the ontrain wifi on this train is even worse than my tethered 2G connection :P</li>
 <li id="l-353"><a href="#l-353">#</a> [18:04] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-354" _a="gsnedders"><a href="#l-354">#</a> [18:06] &lt;gsnedders&gt; Given last time I tried Criic on an unreliable connection it was¬ awkward. Stuff didn't work well. :)</li>
 <li id="l-355" _a="gsnedders"><a href="#l-355">#</a> [18:07] &lt;gsnedders&gt; Also only 15 minutes free</li>
 <li id="l-356" _a="sangwhan" _r="gsnedders"><a href="#l-356">#</a> [18:08] &lt;sangwhan&gt; gsnedders: yeah, critic sucks on a unstable connection..</li>
 <li id="l-357"><a href="#l-357">#</a> [18:11] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-358" _a="odinho"><a href="#l-358">#</a> [18:13] &lt;odinho&gt; Helps turning on async xhr, for percieved fastness. But might make stuff more flaky ;9</li>
 <li id="l-359" _a="sangwhan"><a href="#l-359">#</a> [18:14] &lt;sangwhan&gt; need a r on this <a href="https://critic.hoppipolla.co.uk/r/1107">https://critic.hoppipolla.co.uk/r/1107</a></li>
 <li id="l-360" _a="gsnedders" _r="odinho"><a href="#l-360">#</a> [18:14] &lt;gsnedders&gt; odinho: The problem is sometimes it never times out</li>
 <li id="l-361" _a="odinho"><a href="#l-361">#</a> [18:14] &lt;odinho&gt; It should probably store its data in IndexedDB and sync lators.</li>
 <li id="l-362" _a="sangwhan"><a href="#l-362">#</a> [18:14] &lt;sangwhan&gt; (been stuck in the queue for a long time)</li>
 <li id="l-363" _a="gsnedders" _r="odinho"><a href="#l-363">#</a> [18:14] &lt;gsnedders&gt; odinho: So it doesn't even have a chance to fail and let you retry :(</li>
 <li id="l-364" _a="jgraham" _r="odinho"><a href="#l-364">#</a> [18:17] &lt;jgraham&gt; odinho: Patches welcome ;)</li>
 <li id="l-365" _a="sangwhan" _r="odinho"><a href="#l-365">#</a> [18:18] &lt;sangwhan&gt; odinho: you have access to the real wishlist for critic :)</li>
 <li id="l-366" _a="odinho"><a href="#l-366">#</a> [18:22] &lt;odinho&gt; Why is everyone saying patches welcome to me, when it was gsnedders and sangwhan who were talking? :P</li>
 <li id="l-367" _a="gsnedders" _r="odinho"><a href="#l-367">#</a> [18:25] &lt;gsnedders&gt; odinho: I was just bitchig, you gave concrete suggestions :)</li>
 <li id="l-368" _a="sangwhan" _r="odinho"><a href="#l-368">#</a> [18:26] &lt;sangwhan&gt; odinho: i was actually suggesting if you have a concrete suggestion make jl aware of it :)</li>
 <li id="l-369" _a="sangwhan"><a href="#l-369">#</a> [18:27] * sangwhan was also just ducking out of usual reflex</li>
 <li id="l-370" _a="gsnedders"><a href="#l-370">#</a> [18:27] &lt;gsnedders&gt; :)</li>
 <li id="l-371"><a href="#l-371">#</a> [18:27] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-372"><a href="#l-372">#</a> [18:35] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-373"><a href="#l-373">#</a> [18:37] * Joins: edoyle (~edoyle@public.cloak)</li>
 <li id="l-374"><a href="#l-374">#</a> [18:39] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-375" _a="sangwhan" _r="gsnedders"><a href="#l-375">#</a> [18:45] &lt;sangwhan&gt; gsnedders: so what was your suggestion about getting a coverage report?</li>
 <li id="l-376" _a="sangwhan"><a href="#l-376">#</a> [18:45] &lt;sangwhan&gt; right now the most reliable way i found was to just ack in the source tree, which is well &quot;functional&quot;</li>
 <li id="l-377" _a="Ms2ger" _r="sangwhan"><a href="#l-377">#</a> [18:46] &lt;Ms2ger&gt; sangwhan, &quot;don't&quot;</li>
 <li id="l-378" _a="sangwhan" _r="Ms2ger"><a href="#l-378">#</a> [18:46] &lt;sangwhan&gt; Ms2ger: nah, won't take that as a answer</li>
 <li id="l-379" _a="jgraham" _r="sangwhan"><a href="#l-379">#</a> [18:48] &lt;jgraham&gt; sangwhan: For what purpose?</li>
 <li id="l-380" _a="sangwhan" _r="jgraham"><a href="#l-380">#</a> [18:50] &lt;sangwhan&gt; jgraham: mostly to have people stop in case they are trying to write duplicate tests, finding things in the tree isn't straightforward</li>
 <li id="l-381" _a="sangwhan" _r="jgraham"><a href="#l-381">#</a> [18:52] &lt;sangwhan&gt; jgraham: and to be able to see which subtrees are lacking coverage</li>
 <li id="l-382" _a="sangwhan" _r="sadly"><a href="#l-382">#</a> [18:53] &lt;sangwhan&gt; sadly, i don't have any good ideas. i guess i'm looking for something like a report from emma, as a start</li>
 <li id="l-383" _a="sangwhan" _r="plh"><a href="#l-383">#</a> [19:01] &lt;sangwhan&gt; plh: thanks</li>
 <li id="l-384" _a="plh"><a href="#l-384">#</a> [19:03] &lt;plh&gt; I'm not done yet :)</li>
 <li id="l-385" _a="jgraham" _r="sangwhan"><a href="#l-385">#</a> [19:04] &lt;jgraham&gt; sangwhan: Well that's a super-hard problem</li>
 <li id="l-386" _a="Ms2ger"><a href="#l-386">#</a> [19:04] &lt;Ms2ger&gt; Nah</li>
 <li id="l-387" _a="jgraham"><a href="#l-387">#</a> [19:04] &lt;jgraham&gt; But I think eventually code coverage data will provide that kind of data</li>
 <li id="l-388" _a="Ms2ger"><a href="#l-388">#</a> [19:04] &lt;Ms2ger&gt; It only needs an instrumented reference implementation</li>
 <li id="l-389" _a="jgraham"><a href="#l-389">#</a> [19:04] &lt;jgraham&gt; Too much data</li>
 <li id="l-390"><a href="#l-390">#</a> [19:16] * Quits: ArtB (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-391" _a="sangwhan" _r="plh"><a href="#l-391">#</a> [19:23] &lt;sangwhan&gt; plh: take your time, i'm used to reviews with more nits so that's nothing :)</li>
 <li id="l-392" _a="sangwhan" _r="jgraham"><a href="#l-392">#</a> [19:24] &lt;sangwhan&gt; jgraham: best idea i have is to abuse the devtools/profiler, but that will make it implementation dependent</li>
 <li id="l-393"><a href="#l-393">#</a> [19:28] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-394"><a href="#l-394">#</a> [19:35] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-395"><a href="#l-395">#</a> [19:40] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-396" _a="abarsto"><a href="#l-396">#</a> [19:40] * abarsto is now known as ArtB</li>
 <li id="l-397" _a="mdyck"><a href="#l-397">#</a> [19:50] &lt;mdyck&gt; is anyone working on a reference implementation?</li>
 <li id="l-398" _a="jgraham"><a href="#l-398">#</a> [19:50] &lt;jgraham&gt; Reference implementation of what?</li>
 <li id="l-399" _a="Ms2ger"><a href="#l-399">#</a> [19:50] &lt;Ms2ger&gt; Servo? :)</li>
 <li id="l-400" _a="plh"><a href="#l-400">#</a> [19:50] &lt;plh&gt; lots of browser vendors are working on their reference implementation :)</li>
 <li id="l-401" _a="mdyck"><a href="#l-401">#</a> [19:52] &lt;mdyck&gt; not sure, but i thought sangwhan was talking about html5</li>
 <li id="l-402" _a="Ms2ger" _r="jgraham"><a href="#l-402">#</a> [19:55] &lt;Ms2ger&gt; jgraham, so, you got a good way of catching exceptions in event handlers in subframes?</li>
 <li id="l-403" _a="jgraham" _r="Ms2ger"><a href="#l-403">#</a> [20:00] &lt;jgraham&gt; Ms2ger: Example?</li>
 <li id="l-404" _a="Ms2ger" _r="jgraham"><a href="#l-404">#</a> [20:00] &lt;Ms2ger&gt; jgraham, <a href="https://pastebin.mozilla.org/4743184">https://pastebin.mozilla.org/4743184</a></li>
 <li id="l-405" _a="Ms2ger" _r="jgraham"><a href="#l-405">#</a> [20:00] &lt;Ms2ger&gt; jgraham, (re: bz/ehsan in #content)</li>
 <li id="l-406" _a="jgraham" _r="Ah"><a href="#l-406">#</a> [20:03] &lt;jgraham&gt; Ah, forgot to join #content</li>
 <li id="l-407"><a href="#l-407">#</a> [20:21] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-408"><a href="#l-408">#</a> [20:29] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-409"><a href="#l-409">#</a> [20:35] * Quits: glenn (~gadams@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-410"><a href="#l-410">#</a> [20:36] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-411"><a href="#l-411">#</a> [20:37] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-412"><a href="#l-412">#</a> [20:38] * Quits: anssik (~uid10742@public.cloak) (&quot;Connection closed for inactivity&quot;)</li>
 <li id="l-413"><a href="#l-413">#</a> [21:17] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-414"><a href="#l-414">#</a> [21:30] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-415"><a href="#l-415">#</a> [21:37] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-416"><a href="#l-416">#</a> [21:51] * Joins: shepazu (schepers@public.cloak)</li>
 <li id="l-417"><a href="#l-417">#</a> [22:13] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-418"><a href="#l-418">#</a> [22:17] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-419"><a href="#l-419">#</a> [22:31] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-420"><a href="#l-420">#</a> [22:33] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-421"><a href="#l-421">#</a> [22:38] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-422" _a="heycam|away"><a href="#l-422">#</a> [23:01] * heycam|away is now known as heycam</li>
 <li id="l-423"><a href="#l-423">#</a> [23:16] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-424"><a href="#l-424">#</a> [23:16] -gitbot:#testing- [web-platform-tests] dontcallmedom pushed 3 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/07e7ef04288b...c47c4fee62cf">https://github.com/w3c/web-platform-tests/compare/07e7ef04288b...c47c4fee62cf</a></li>
 <li id="l-425"><a href="#l-425">#</a> [23:16] -gitbot:#testing- web-platform-tests/master 56860c1 Dominique Hazael-Massieux: updated to use common vendor prefix infrastructure</li>
 <li id="l-426"><a href="#l-426">#</a> [23:16] -gitbot:#testing- web-platform-tests/master 8e6d632 Dominique Hazael-Massieux: removing extraneous white space</li>
 <li id="l-427"><a href="#l-427">#</a> [23:16] -gitbot:#testing- web-platform-tests/master c47c4fe Dominique Hazael-Massieux: Merge pull request #800 from dontcallmedom/gum-prefix...</li>
 <li id="l-428"><a href="#l-428">#</a> [23:16] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-429"><a href="#l-429">#</a> [23:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-430"><a href="#l-430">#</a> [23:22] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/single_page (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/0d556f3f0c61">https://github.com/w3c/web-platform-tests/commit/0d556f3f0c61</a></li>
 <li id="l-431"><a href="#l-431">#</a> [23:22] -gitbot:#testing- web-platform-tests/jgraham/single_page 0d556f3 James Graham: Use single_page branch of testharness.js</li>
 <li id="l-432"><a href="#l-432">#</a> [23:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-433"><a href="#l-433">#</a> [23:23] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-434"><a href="#l-434">#</a> [23:23] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #829: Use single_page branch of testharness.js (master...jgraham/single_page) <a href="https://github.com/w3c/web-platform-tests/pull/829">https://github.com/w3c/web-platform-tests/pull/829</a></li>
 <li id="l-435"><a href="#l-435">#</a> [23:23] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-436" _a="jgraham"><a href="#l-436">#</a> [23:25] &lt;jgraham&gt; Dammit the sync script does sync /resources</li>
 <li id="l-437" _a="jgraham"><a href="#l-437">#</a> [23:25] &lt;jgraham&gt; *doesn</li>
 <li id="l-438" _a="jgraham"><a href="#l-438">#</a> [23:25] &lt;jgraham&gt; 't</li>
 <li id="l-439" _a="jgraham"><a href="#l-439">#</a> [23:25] &lt;jgraham&gt; There goes my evil plan to abuse w3c-test.org to demo some testharness.js stuff</li>
 <li id="l-440" _a="jgraham"><a href="#l-440">#</a> [23:26] &lt;jgraham&gt; Or mayve I forgot to push…</li>
 <li id="l-441"><a href="#l-441">#</a> [23:31] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-442"><a href="#l-442">#</a> [23:39] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-443"><a href="#l-443">#</a> [23:41] * Quits: ptressel (~chatzilla@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-444"><a href="#l-444">#</a> Session Close: Wed Apr 02 00:00:01 2014</li>
</ol>
<p id="end">The end :)
